+++ jax/_src/__init__.py
@@ -11,10 +11,3 @@
 # WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 # See the License for the specific language governing permissions and
 # limitations under the License.
-
-# flake8: noqa: F401
-
-from jax._src.scipy.stats.geom import (
-  logpmf,
-  pmf,
-)
