+++ jax/_src/scipy/optimize/line_search.py
@@ -11,6 +11,7 @@
 # WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 # See the License for the specific language governing permissions and
 # limitations under the License.
+
 from typing import NamedTuple, Union
 
 import jax.numpy as jnp
