+++ jax/scipy/stats/bernoulli.py
@@ -1,4 +1,4 @@
-# Copyright 2018 Google LLC
+# Copyright 2020 Google LLC
 #
 # Licensed under the Apache License, Version 2.0 (the "License");
 # you may not use this file except in compliance with the License.
@@ -12,25 +12,9 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
+# flake8: noqa: F401
 
-import scipy.stats as osp_stats
-
-from ... import lax
-from ...numpy import lax_numpy as jnp
-from ...numpy._util import _wraps
-from ..special import xlogy, xlog1py
-
-
-@_wraps(osp_stats.bernoulli.logpmf, update_doc=False)
-def logpmf(k, p, loc=0):
-  k, p, loc = jnp._promote_args_inexact("bernoulli.logpmf", k, p, loc)
-  zero = jnp._constant_like(k, 0)
-  one = jnp._constant_like(k, 1)
-  x = lax.sub(k, loc)
-  log_probs = xlogy(x, p) + xlog1py(lax.sub(one, x), -p)
-  return jnp.where(jnp.logical_or(lax.lt(x, zero), lax.gt(x, one)),
-                  -jnp.inf, log_probs)
-
-@_wraps(osp_stats.bernoulli.pmf, update_doc=False)
-def pmf(k, p, loc=0):
-  return jnp.exp(logpmf(k, p, loc))
+from jax._src.scipy.stats.bernoulli import (
+  logpmf,
+  pmf,
+)
