+++ jax/scipy/stats/cauchy.py
@@ -1,4 +1,4 @@
-# Copyright 2018 Google LLC
+# Copyright 2020 Google LLC
 #
 # Licensed under the Apache License, Version 2.0 (the "License");
 # you may not use this file except in compliance with the License.
@@ -12,23 +12,9 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
+# flake8: noqa: F401
 
-import numpy as np
-import scipy.stats as osp_stats
-
-from ... import lax
-from ...numpy._util import _wraps
-from ...numpy.lax_numpy import _promote_args_inexact, _constant_like
-
-
-@_wraps(osp_stats.cauchy.logpdf, update_doc=False)
-def logpdf(x, loc=0, scale=1):
-  x, loc, scale = _promote_args_inexact("cauchy.logpdf", x, loc, scale)
-  pi = _constant_like(x, np.pi)
-  scaled_x = lax.div(lax.sub(x, loc), scale)
-  normalize_term = lax.log(lax.mul(pi, scale))
-  return lax.neg(lax.add(normalize_term, lax.log1p(lax.mul(scaled_x, scaled_x))))
-
-@_wraps(osp_stats.cauchy.pdf, update_doc=False)
-def pdf(x, loc=0, scale=1):
-  return lax.exp(logpdf(x, loc, scale))
+from jax._src.scipy.stats.cauchy import (
+  logpdf,
+  pdf,
+)
