+++ jax/scipy/stats/logistic.py
@@ -12,33 +12,13 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
-import scipy.stats as osp_stats
-from jax.scipy.special import expit, logit
+# flake8: noqa: F401
 
-from ... import lax
-from ...numpy._util import _wraps
-
-
-@_wraps(osp_stats.logistic.logpdf, update_doc=False)
-def logpdf(x):
-  return lax.neg(x) - 2. * lax.log1p(lax.exp(lax.neg(x)))
-
-@_wraps(osp_stats.logistic.pdf, update_doc=False)
-def pdf(x):
-  return lax.exp(logpdf(x))
-
-@_wraps(osp_stats.logistic.ppf, update_doc=False)
-def ppf(x):
-  return logit(x)
-
-@_wraps(osp_stats.logistic.sf, update_doc=False)
-def sf(x):
-  return expit(lax.neg(x))
-
-@_wraps(osp_stats.logistic.isf, update_doc=False)
-def isf(x):
-  return -logit(x)
-
-@_wraps(osp_stats.logistic.cdf, update_doc=False)
-def cdf(x):
-  return expit(x)
+from jax._src.scipy.stats.logistic import (
+  cdf,
+  isf,
+  logpdf,
+  pdf,
+  ppf,
+  sf,
+)
