+++ jax/scipy/stats/norm.py
@@ -1,4 +1,4 @@
-# Copyright 2018 Google LLC
+# Copyright 2020 Google LLC
 #
 # Licensed under the Apache License, Version 2.0 (the "License");
 # you may not use this file except in compliance with the License.
@@ -12,43 +12,12 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
+# flake8: noqa: F401
 
-import numpy as np
-import scipy.stats as osp_stats
-
-from ... import lax
-from ...numpy import lax_numpy as jnp
-from ...numpy._util import _wraps
-from ...numpy.lax_numpy import _promote_args_inexact, _constant_like
-from .. import special
-
-@_wraps(osp_stats.norm.logpdf, update_doc=False)
-def logpdf(x, loc=0, scale=1):
-  x, loc, scale = _promote_args_inexact("norm.logpdf", x, loc, scale)
-  two = _constant_like(x, 2)
-  scale_sqrd = lax.pow(scale, two)
-  log_normalizer = lax.log(lax.mul(_constant_like(x, 2 * np.pi), scale_sqrd))
-  quadratic = lax.div(lax.pow(lax.sub(x, loc), two), scale_sqrd)
-  return lax.div(lax.neg(lax.add(log_normalizer, quadratic)), two)
-
-
-@_wraps(osp_stats.norm.pdf, update_doc=False)
-def pdf(x, loc=0, scale=1):
-  return lax.exp(logpdf(x, loc, scale))
-
-
-@_wraps(osp_stats.norm.cdf, update_doc=False)
-def cdf(x, loc=0, scale=1):
-  x, loc, scale = _promote_args_inexact("norm.cdf", x, loc, scale)
-  return special.ndtr(lax.div(lax.sub(x, loc), scale))
-
-
-@_wraps(osp_stats.norm.logcdf, update_doc=False)
-def logcdf(x, loc=0, scale=1):
-  x, loc, scale = _promote_args_inexact("norm.logcdf", x, loc, scale)
-  return special.log_ndtr(lax.div(lax.sub(x, loc), scale))
-
-
-@_wraps(osp_stats.norm.ppf, update_doc=False)
-def ppf(q, loc=0, scale=1):
-  return jnp.array(special.ndtri(q) * scale + loc, 'float64')
+from jax._src.scipy.stats.norm import (
+  cdf,
+  logcdf,
+  logpdf,
+  pdf,
+  ppf,
+)
