+++ jax/scipy/stats/pareto.py
@@ -1,4 +1,4 @@
-# Copyright 2018 Google LLC
+# Copyright 2020 Google LLC
 #
 # Licensed under the Apache License, Version 2.0 (the "License");
 # you may not use this file except in compliance with the License.
@@ -12,23 +12,9 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
+# flake8: noqa: F401
 
-import scipy.stats as osp_stats
-
-from ... import lax
-from ...numpy._util import _wraps
-from ...numpy.lax_numpy import _promote_args_inexact, _constant_like, inf, where
-
-
-@_wraps(osp_stats.pareto.logpdf, update_doc=False)
-def logpdf(x, b, loc=0, scale=1):
-  x, b, loc, scale = _promote_args_inexact("pareto.logpdf", x, b, loc, scale)
-  one = _constant_like(x, 1)
-  scaled_x = lax.div(lax.sub(x, loc), scale)
-  normalize_term = lax.log(lax.div(scale, b))
-  log_probs = lax.neg(lax.add(normalize_term, lax.mul(lax.add(b, one), lax.log(scaled_x))))
-  return where(lax.lt(x, lax.add(loc, scale)), -inf, log_probs)
-
-@_wraps(osp_stats.pareto.pdf, update_doc=False)
-def pdf(x, b, loc=0, scale=1):
-  return lax.exp(logpdf(x, b, loc, scale))
+from jax._src.scipy.stats.pareto import (
+  logpdf,
+  pdf,
+)
