+++ jax/scipy/stats/t.py
@@ -1,4 +1,4 @@
-# Copyright 2018 Google LLC
+# Copyright 2020 Google LLC
 #
 # Licensed under the Apache License, Version 2.0 (the "License");
 # you may not use this file except in compliance with the License.
@@ -12,29 +12,9 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
+# flake8: noqa: F401
 
-import numpy as np
-import scipy.stats as osp_stats
-
-from ... import lax
-from ...numpy._util import _wraps
-from ...numpy.lax_numpy import _promote_args_inexact, _constant_like
-
-
-@_wraps(osp_stats.t.logpdf, update_doc=False)
-def logpdf(x, df, loc=0, scale=1):
-  x, df, loc, scale = _promote_args_inexact("t.logpdf", x, df, loc, scale)
-  two = _constant_like(x, 2)
-  scaled_x = lax.div(lax.sub(x, loc), scale)
-  df_over_two = lax.div(df, two)
-  df_plus_one_over_two = lax.add(df_over_two, _constant_like(x, 0.5))
-  normalize_term_const = lax.mul(lax.mul(scale, scale), _constant_like(x, np.pi))
-  normalize_term_tmp = lax.div(lax.log(lax.mul(normalize_term_const, df)), two)
-  normalize_term = lax.sub(lax.add(lax.lgamma(df_over_two), normalize_term_tmp),
-                           lax.lgamma(df_plus_one_over_two))
-  quadratic = lax.div(lax.mul(scaled_x, scaled_x), df)
-  return lax.neg(lax.add(normalize_term, lax.mul(df_plus_one_over_two, lax.log1p(quadratic))))
-
-@_wraps(osp_stats.t.pdf, update_doc=False)
-def pdf(x, df, loc=0, scale=1):
-  return lax.exp(logpdf(x, df, loc, scale))
+from jax._src.scipy.stats.t import (
+  logpdf,
+  pdf,
+)
